# docs/swagger.yaml
definitions:
...
  internal_controllers.CreatePermissionRequest:
    properties:
      action:
        maxLength: 50
        minLength: 2
        type: string
      description:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      resource:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - action
    - name
    - resource
    type: object
...
  internal_controllers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_controllers.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  internal_controllers.ValidateTokenRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  user-management_internal_utils.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      summary: Forgot password
      tags:
      - Password Reset
...
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      summary: Reset password
      tags:
      - Password Reset
  /api/v1/auth/validate-reset-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.ValidateTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      summary: Validate reset token
      tags:
      - Password Reset
...
  /api/v1/permissions:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all permissions
      tags:
      - Permissions
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Permission data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.CreatePermissionRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Create permission
      tags:
      - Permissions
...
  /api/v1/roles/{roleId}/permissions:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get permissions by role
      tags:
      - Permissions
  /api/v1/roles/{roleId}/permissions/{permissionId}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      - description: Permission ID
        in: path
        name: permissionId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove permission from role
      tags:
      - Permissions
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      - description: Permission ID
        in: path
        name: permissionId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Assign permission to role
      tags:
      - Permissions
...
swagger: "2.0"